Add search_by_bugs() method and sample files.
The search_by_bugs() method is the main control function responsible to look on the "landscape" file for the first line of the "bug" file, indicating as a potential bug in the content. Once found, do calculations to look for the exactly number of next lines of the landscape content as the same of the total of the bug lines, and the same position on the line (column). If found the same patterns (bug[line] == pattern[next_line]), we have a bug found.

Signed-off-by: Paulo Vital <pvital@gmail.com>